requirements.txt
streamlit>=1.20.0
python-dotenv>=1.0.0
google-generativeai>=0.3.1
requests>=2.31.0
